fix some more imports
