+++ airbyte-ci/connectors/pipelines/.gitignore
@@ -1,1 +1,2 @@
 pipeline_reports
+.venv
