+++ airbyte-ci/connectors/pipelines/pipelines/pipeline/connectors/context.py
@@ -15,7 +15,7 @@ from dagger import Directory
 from github import PullRequest
 from pipelines.dagger.actions import secrets
 from pipelines.models.bases import ConnectorReport, ConnectorWithModifiedFiles
-from pipelines.models.context import PipelineContext
+from pipelines.models.contexts import PipelineContext
 from pipelines.helpers.github import update_commit_status_check
 from pipelines.helpers.slack import send_message_to_webhook
 from pipelines.helpers.utils import METADATA_FILE_NAME
